Review pass: fix a migration-fatal type change and an upsert batch collision
Two real defects found reviewing the ingest commit before merge:

1. The migration would have FAILED on apply. The live view's
   installs/trial_starts/purchases columns are bigint (sum over integer).
   The restructured meta arm pre-aggregates in a CTE, so the outer query
   summed bigint -- and sum(bigint) returns numeric. create-or-replace
   refuses to change an existing view column's type, so db-migrate.yml
   would have errored on merge to prod. The earlier 381-row EXCEPT
   validation compared values, not types, which is exactly the blind spot.
   Fixed with explicit ::bigint casts; the full view body's output types
   now verified against prod pg_attribute (numeric, bigint, bigint,
   bigint, numeric + appended text).

2. The first live upsert with real data would have 500d. The report
   requests `app` and `os` as dimensions, so Singular returns one row per
   (app, os, campaign, date) -- multiple rows sharing the table's
   (source, campaign_id, date) PK. Duplicate conflict keys in a single
   PostgREST INSERT make Postgres reject the whole batch ("ON CONFLICT DO
   UPDATE command cannot affect row a second time"). mapSingularRows now
   merges to PK grain: metrics summed, os kept only when every fanned row
   agrees on it, `app` consumed so it never shows up as unmapped noise.

Also corrects the cohort-catalog cache comment: it lives for the process
lifetime (many cron invocations on a warm lambda), not the request --
acceptable staleness, but the comment shouldn't lie about it.

17 vendor tests pass (+2 for the merge behavior); tsc --noEmit clean.

+++ supabase/migrations/0065_singular_attribution.sql
@@ -131,9 +131,14 @@ select
   max(campaign_name)  as campaign_name,
   date,
   sum(spend)          as spend,
-  sum(installs)       as installs,
-  sum(trial_starts)   as trial_starts,
-  sum(purchases)      as purchases,
+  -- ::bigint casts are load-bearing. The live view's integer-ish columns are
+  -- bigint (sum over integer). The meta arm here is PRE-aggregated, so the
+  -- outer sum runs over bigint, and sum(bigint) returns NUMERIC — and
+  -- create-or-replace refuses to change an existing column's type, so without
+  -- the casts this migration fails on apply. Verified against prod pg_attribute.
+  sum(installs)::bigint       as installs,
+  sum(trial_starts)::bigint   as trial_starts,
+  sum(purchases)::bigint      as purchases,
   sum(purchase_value) as purchase_value,
   max(trial_source)   as trial_source
 from unioned
